drm: xilinx: dp_sub: Program the scalining factor based on the format
commit 443c13510738baa548796d2b715c8e047fa29896 from
https://github.com/Xilinx/linux-xlnx.git

The scaling factor should be programmed as the format changes. Modify
the xilinx_drm_dp_sub_av_buf_init_sf() so that the function can
selectively program the scaling factor for each layer. The function
can be used to update the scaling factor when the layer format changes
in xilinx_drm_dp_sub_layer_set_fmt();

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>