ASoC: Intel: Skylake: Increase the max number of entries
commit 821993586259fb46e79ca7d2eb32d2a79b0d8df8 from
git://git@github.com:01org/linux-intel-4.9.git
for resources, path configs and formats

To support multi-format playback/capture, 0.7 xml has more
number of entries for module resources, path configs and formats.

This patch increases the max limit for module resources to 32,
path config entries to 32 and module formats to 64. Although
this is not a permanent solution. Right solution is to dynamically
calculate the entries instead of hard coding.

Change-Id: If052e6b95a69b9ed47f08ebe284383c4d2e8e81f
Signed-off-by: Puneeth Prabhu <puneethx.prabhu@intel.com>
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/14286
Reviewed-by: audio_build <audio_build@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Reviewed-by: Kale, Sanyog R <sanyog.r.kale@intel.com>
Reviewed-by: Singh, Guneshwor O <guneshwor.o.singh@intel.com>
Reviewed-by: R, Dharageswari <dharageswari.r@intel.com>
Reviewed-by: Nc, Shreyas <shreyas.nc@intel.com>
Reviewed-by: Diwakar, Praveen <praveen.diwakar@intel.com>
Tested-by: Sm, Bhadur A <bhadur.a.sm@intel.com>