clk: zynqmp: Let the frac be decided on the frac capability
commit 5d029fdc257cf88e65500db348eda23040af332b from
https://github.com/Xilinx/linux-xlnx.git

Currently if the set_rate_parent is set and there is  remainder the
second divisor is made to one. This works well if the parent is fractional
or can change rate.
In case the tree is diva --> divb--> fixed_parent

the expectation is both diva and divb will get us to the desired
frequency. However if the calculation has a remainder then diva is
always 1 and in some cases saturates divb.

Add a flag CLK_FRAC to check and only then make the div as 1.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
Acked-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>