remoteproc: add a rproc_set_firmware() API
A new API, rproc_set_firmware() is added to allow the remoteproc client
drivers to be able to configure a custom firmware name that is different
from the default name registered by a remoteproc platform driver. This
function is being introduced to provide a kernel-level equivalent of
the current sysfs interface, so as to allow these kernel-drivers to
choose different firmwares based on their functional feature it is
implementing using the remote processor.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Jacob Stiffler <j-stiffler@ti.com>
(cherry picked from commit 3686765450158be2e80a80240d02a18473e9b23e)