arm64: zynqmp: Add support for Xilinx zc1232 board
commit 7e6efd96a1f7bd9021ae84d193e2103869e20834 from
https://github.com/Xilinx/linux-xlnx.git

This patch adds support for Xilinx zc1232 board.

Only QSPI(single) and uarts are wired on this board.
Sata is enabled based on request because it is connected to
Bullseye headers (4-lane GT cable header to SMA cables) where
a SMA-to-SATA adapter can be connected to get sata to work.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>