i2c: cadence: Fixed repeated start not holding the bus long enough
commit 9e90cc17df450120bbb110e83e8a66bba9e36855 from
https://github.com/Xilinx/linux-xlnx.git

The driver was clearing the hold bit in the control register before
writing to the address register so that a stop condition was
generated rather than a repeated start.

This issue was only when the system was running much slower than
normal such that it is not typically seen.  The IP data sheet
says to write to the address register before clearing the hold
bit generally.

Signed-off-by: John Linn <john.linn@xilinx.com>
Tested-by: Nava kishore Manne <navam@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>