drm: xilinx: plane: improve handling of plane dma channels
commit 03066e43dec534ce7a23f37d52262ddf7914e01e from
https://github.com/Xilinx/linux-xlnx.git

With the previous code, it was possible to exit the function without
having any dma channels enabled if they all failed with -ENODEV. Failure
of the primary plane would break a hard assumption and cause a null
pointer dereference further in xilinx_drm_plane_get_align.

Instead of trying each dma-name, then trying to guess if it existed or
not in the DT, loop over the names that are actually given in the DT.
This allows to properly fail if a dma-name is not found, signaling a bad
plane node setup in DT.

Also improve the error feedback to help developers properly identify the
DT node setup error.

Signed-off-by: Jean-Francois Dagenais <jeff.dagenais@gmail.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>