Invalidate local tlb before cpu down
The local TLB should have been invalidated in cpu hotplug to avoid a scenario
that when a pluged cpu came back(online) and processes being migrated onto it,
there is a chance that one of their PIDs happens to match the PID field of a
entry in TLB, then a segfault would happen to the process.

This could be verified with taking a test, like below:
 # while :;do ls > /dev/null; done &
 # while :; do \
   for i in `seq 1 ${MAX_CPU}`; do echo 0 > /sys/devices/system/cpu/cpu$i/online;done; \
   for i in `seq 1 ${MAX_CPU}`; do echo 1 > /sys/devices/system/cpu/cpu$i/online; done;\
   done &

In several minutes or more, a segfault will happen to 'ls', like below:

 # Cannot set affinity for irq 17
 sh[5439]: unhandled signal 11 at 00000000 nip 00000000 lr 1002edac code 30001

 [1]-  Segmentation fault      while :; do
     ls > /dev/null;
 done

To fix this issue, we invalidate the local TLB before a CPU down.

Signed-off-by: Jianchao Wang <jianchao.wang@windriver.com>
Signed-off-by: Qi Hou <qi.hou@windriver.com>