dma: xilinx: New driver for Video Framebuffer IP
commit 166f3acd1fdde298178852d6db7c29ce80b13d76 from
https://github.com/Xilinx/linux-xlnx.git

The Video Framebuffer IP is available in two forms: read or write.
This driver supports either form of the IP.
Each IP is a single channel DMA but which is video format aware
supporting video data that arranged in memory as packed or semi-planar data.
This driver introduces support for these IP and integrates with the
DMA Engine framework.

Signed-off-by: Radhey Pandey <radheys@xilinx.com>
Signed-off-by: John Nichols <jnichol@xilinx.com>
Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>