drm: xilinx: Enable use of Framebuffer Read IP in Xilinx DRM Driver
commit 37f313ae56790a64990f25181a1fb7672b9a62a1 from
https://github.com/Xilinx/linux-xlnx.git

Xilinx Video Framebuffer read is a 'format aware' DMA device.  It
requires knowledge of the video memory format to be explicitly
programmed by DMA clients.  This change enables this data to be
passed via the dma channel private data pointer as there is no
standard DMA Engine interface for passing this information.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>