media: xilinx: dma: Enable use of Framebuffer Write IP in V4L2 Driver
commit d3ac87f034613cadc97cb34b11dd386b1039cdaf from
https://github.com/Xilinx/linux-xlnx.git

The Xilinx Video Framebuffer driver is video 'format aware' and
requires clients to pass video format information as part of DMA
programming. As the DMA Engine interface does not support passing
this information, a configuration object is passed through the
dma channel's private data to the Framebuffer Driver with this
information.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>