dwc3: Add support for clock disabling during suspend
commit 80ac0da9010ebabef6ed1d66b6607b345fd14cb4 from
https://github.com/Xilinx/linux-xlnx.git

In dwc3 usb controller, when remote wakeup feature is enabled and
the controller is in U3 state (suspended), it requires suspend_clk
to be active to detect wakeup signalling driven on the link.
For all other cases usb clocks can be disabled once the controller
enters U3 state.

This patch adds the logic for disabling usb clocks when remote
wakeup feature is not enabled.

Signed-off-by: Anurag Kumar Vulisha <anuragku@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>