arm64: zynqmp: Label zcu106 with board revision
commit e9ac9e496c0f93e59c93a9730232e718d62709d9 from
https://github.com/Xilinx/linux-xlnx.git

Not using board revision is causing confusion about which board is
supported and tested. Mark dts files exactly with board revision which was
tested. When new board revision arives it can be symlink if SW view is
the same. Also add -revX suffix to compatible string because user space
tools are parsing this string and can change behavior depends of board
revision.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>