net: ethernet: xilinx: Fix race condition in the tx path
commit ee40c7e77f9b3935b1b88b19345644cec1e1242b from
https://github.com/Xilinx/linux-xlnx.git

In the _xmit driver is checking for the buffer descriptor
availabilty, If the buffer descriptors are not available
it is stopping all the device queues,
It should stop only that particular queue as
other queues still may have free buffer descriptors
for servicing the data.

This patch fixes this issue.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Acked-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>