drm: xilinx: crtc: Fix the pixel clock enablement
commit 955a80b3e5b2a30e42524b4cd9d2d37014be3830 from
https://github.com/Xilinx/linux-xlnx.git

The pixel clock doesn't have to be enabled in probe. Enable
the clock upon DPMS on request. The clock needs to be enabled
first to make sure that CCF disables the clock as CCF skips
the disable request for already disabled clock.
And the clock enable flag was always set to true upon DPMA on.
Fix it to be set to false when the clk_prepare_enable() fails.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>