soc: zynqmp: Extract firmware part from PM
commit 03072221332381377af031308cf840a7e69db17e from
https://github.com/Xilinx/linux-xlnx.git

The purpose of this patch is to separate code used by drivers for
communication with firmware and power management functions.

New Kconfig option is added for enabling ZYNQMP_PM to enable power
management only. Low level wiring is enabled when SOC_XILINX_ZYNQMP is
enabled.

There is only one difference in behavior which is that pm platform
driver is asking for pm_api_version again and not reading this value
from low level driver.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>