v4l: xilinx-vpss-csc: Any-to-any color space conversion
commit 1437167470651ab76267bb0433274efb0e22a758 from
https://github.com/Xilinx/linux-xlnx.git

In 2017.3, the VPSS CSC IP supports any-to-any color space converion.
This implies the color space conversion can occur from an
input format consisting of - RGB, YUV 444, YUV 422, YUV 420 to an
output format list consisting of - RGB, YUV 444, YUV 422, YUV 420.

The IP supports YUV 420, but due to the lack of YUV 420 in
media-bus-format.h that support in the driver will be deferred to later.

Signed-off-by: Rohit Athavale <rathaval@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>