zynqmp: sdhci: Prevent SD driver deferred probe after host init
commit a49e5e4ac91e660a2b58f845b8588cf9ae3837bf from
https://github.com/Xilinx/linux-xlnx.git

This patch prevents deferred probe after the host initialization is
done. Instead if the deferred probe is necessary, do it before the host
initialization. This way the unnecessary increase in host id can be
avoided.
In standard boot flow deferred probe is called at least 2 times
that's why it is better to extract it out.
Also add error message when getting silicon version fails.

Signed-off-by: Manish Narani <mnarani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>