phy-ns2-usb2: USB PHY sleep and suspend override
Single PHY is used for EHCI and OHCI controller.
If EHCI controller goes to suspend, PHY also enters to
Suspend. By that time OHCI controller might be still active.
If OHCI Controller tries to enters to suspend, kernel crash
is observed.
we add the settings overwrites the Controller UTMI suspend and
sleep signals going to PHY.
As a result, we can avoid the kernel panic as belows:

  Bad mode in Error handler detected on CPU3, code 0xbf000002 -- SError
  CPU: 3 PID: 362 Comm: rrdcached Not tainted 4.8.18-rt10-WR9.0.0.4_preempt-rt #1
  Hardware name: Broadcom NS2 SVK (DT)
  Call trace:
  [<ffffff8008088318>] dump_backtrace+0x0/0x1d0
  [<ffffff800808850c>] show_stack+0x24/0x30
  [<ffffff800842fe14>] dump_stack+0x90/0xb4
  [<ffffff80080c36d0>] ___might_sleep+0x108/0x170
  [<ffffff80087f7054>] rt_spin_lock+0x2c/0x88
  [<ffffff80080a9bc0>] do_force_sig_info+0x30/0xe0
  [<ffffff80080aa434>] force_sig_info+0x34/0x48
  [<ffffff800808880c>] arm64_notify_die+0x7c/0x90
  [<ffffff8008088c48>] bad_mode+0xb0/0xc8
  [<0000007f919599dc>] 0x7f919599dc
  Internal error: Attempting to execute userspace memory: 8600000f [#1] PREEMPT SMP

There is no the author info of the patch in the LDK tarball.

[cherry-picked from broadcom ns2 LDK 5.0.12 tarball]
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>