driver : phy: trigger to call phy_state_machine() after phy auto-negotiation
On ls1046a-rdb board, there is 10 copper port that uses phy AQR107
with irq mode. As expectation, there should be a interrupt generated,
and then trigger phy_state_machine to be invoked so that change phy
into correct state. But in fact, there is no interrupt generated. So,
in order to let phy enter correct state, it is need do trigger
phy_state_machine to be invoked manually.

Signed-off-by: Meng Li <Meng.Li@windriver.com>