driver : usb: dwc3: there is no need to report a warning trace
In function dwc3_frame_length_adjustment(), if the frame length
value that will be set into register is the same with the value in
register, there will a warning trace is reported. It is a little
making a fuss, so change it to a normal warning and report it with
function dev_warn().

Signed-off-by: Meng Li <Meng.Li@windriver.com>