ARM64: zynqmp: Do not set requirements to 0 for wakeup sources
commit bf761476aa37ffd952748e29ace6eeb4108a1102 from
https://github.com/Xilinx/linux-xlnx.git

Devices which are set as wakeup source or belongs to wakeup
source device's path should not be powered off by generic power
domain driver.

Add check in zynqmp GPD power off function to check if device
is in wakeup source path. If so, set capabilities to WAKEUP
instead of 0 in GPD power off function.

Signed-off-by: Rajan Vaja <rajanv@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>