Documentation: devicetree: bindings: dma: New dts property
commit eed0c52278895ff4f26fd7a0e4f373f76276f540 from
https://github.com/Xilinx/linux-xlnx.git

A new device tree property is described that will describe
the video formats supported in the Video Framebuffer DMA device.
The Video Framebuffer IP is configurabe and can be configured with
varying support for a number of possible video memory formats in
an effort to tailor the size of the logic footprint.  The driver
will utilize this new device tree property to describe this
configuration.

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>