dma: xilinx: Add private API to permit retrieval of supported mem formats
commit f618d3b9e121d8036139419b139f7011903c77dc from
https://github.com/Xilinx/linux-xlnx.git

The video Framebuffer DMA IP requires clients to send a fourcc code
to indicate the memory format layout.  The IP can be configured to support
a variety of memory formats ranging from YUYV, RGB and in either 8 bit
or 10 bit formats.  There has been no method for clients to obtain
this list of supported formats.  This patch adds private APIs that
can be called from clients to retrieve this list so that user space
applications can choose from any of the available memory formats.

Depends on patch 13fd162 (dma: xilinx: Bug fix to ensure only video formats
enabled in IP are in driver)

Signed-off-by: Jeffrey Mouroux <jmouroux@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>