drm: xilinx: dp: Add delay after Monitor wake up
commit f5e4ff5b0f2995fb1bbdb93cdc7e6cecef710bd9 from
https://github.com/Xilinx/linux-xlnx.git

Some monitors require delay to fully wake up. Otherwise, it may
result in some error such as training failure.

Delay of 4 msec was not specified in the spec, but found from
experimentation (ex, no failure for 20 times or more). Thus,
this setting is exposed as module parameter so that user can
change if needed.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>