sched: match on TASK_INTERRUPTIBLE too
The commit [sched: Prevent task state corruption by spurious lock
wakeup] assumed additional changes (specifically in futex_pi) were
present. These were not in the scope of the backport to fix the task
state corruption, however. As such the assumption that all 'sleepers'
would be TASK_UNITERRUPTIBLE was incorrect. It is possible that
'sleepers' are in fact TASK_INTERRUPTIBLE given the current state of
the kernel. We therefor use TASK_NORMAL which will match on both
TASK_UNITERRUPTIBLE and TASK_INTERRUPTIBLE in try_to_wake_up(). Since
we are not reverting to use TASK_ALL we are still excluding
TASK_TRACED and as such the intent of the 'state corruption' patch
remains true and there should be no regression around state
corruption.

Without this change tasks can permamently 'sleep' causing issues like
VMs failing to boot with kvm tasks unable to wake.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>