driver: pci: set correct irq number for pme feature
In init_service_irqs() function, irq for pme is set as dev->irq(pci
device irq), it is not correct. We should use the correct pme irq
number(), so, add "pme" string in property interrupt-names so that
we can get the correct irq number.
In additional, there is redundant operation to initialize pme feature,
so delete it on.

Signed-off-by: Meng Li <Meng.Li@windriver.com>