sound: soc: add audio feature support for codec adau1761
In order to implement audio feature beased on codec adau1761, get file
zed_adau1761.c is from https://github.com/analogdevicesinc/linux.git
The key point of supporting codec adau1761 is to enable FPGA clock.
Because the MCLK of codec adau1761 if from FPGA side.
The firmware for codec adau1761 is used by DSP module that is a part of
adau1761. Without firmware, codec adau1761 can also work well, so it is
not a necessary thing and will not be downloaded on mini-itx platform.
If users have their own firmware for codec adau1761, they can remove
the platform limitation in adau1761.c.

Signed-off-by: Meng Li <Meng.Li@windriver.com>