Revert "sched/core: Free the stack early if CONFIG_THREAD_INFO_IN_TASK"
This reverts commit f585ac3be9123dfe65cf43d3114d935cdf9e6266.
(was commit 68f24b08ee892d47bdef925d676e1ae1ccc316f8 upstream.)

Since this commit was not strictly required to establish the context
needed to apply some of the more complex CVE patches in the last
release, and since it is actually an optimization and not really
a fix, we revert it.

The decision to do this was partly based on finding it caused issues
in conjunction with preempt-rt patches, and historically we have found
that preempt-rt finds issues that are in mainline that are just simply
harder to trigger.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>