pci: fix wrong end calculated by pci_resource_to_user
Some special archs such as mips, ppc and sparc, use a specified
pci_resource_to_user function to calculate the start and the end
of pci resource. But the function has a bug when calculating the
end value. It first gets the resource size by "end - start + 1",
which is all right. Then it uses the return value of fixup_bigphys_addr
as the start of resource, this only matters to Alchemy platforms,
on other platforms fixup_bigphys_addr is just an identidy mapping.
Finally, it sets the end value to "start + size", which is obviously
wrong, because the value becomes "end + 1".

Signed-off-by: Rui Wang <rui.wang@windriver.com>