Revert "sched/core: Enable might_sleep() and smp_processor_id() checks early"
This reverts commit a8a60c655f41da7313f5884010597a9666a9b1e4.

That commit enables might_sleep() and smp_processor_id() checks
right when the scheduler starts working, which is used to find
kernel bug while fixing issue LIN9-6907. But this feature needs
a new system state: SYSTEM_SCHEDULING. And after the new state
was introduced, a regression related to iommu was found. We
finally figure out that we need more patches to make kernel
adjust to the new state, which will be too big for this kind of
issue. So it would be more appropriate to revert the patch.

Signed-off-by: Rui Wang <rui.wang@windriver.com>