Revert "init: Introduce SYSTEM_SCHEDULING state"
This reverts commit 4339a6032394575420daf56604b90282e3026f1a.

That commit introduces a new system state SYSTEM_SCHEDULING, which
is used to support commit a8a60c655f41da7313f5884010597a9666a9b1e4.
But kernel needs more patches to adjust to the new state, which will
be too big for this kind of issue. So it would be more appropriate
to revert the patch.

Signed-off-by: Rui Wang <rui.wang@windriver.com>