thermal: int340x_thermal: Switch to use new generic UUID API
commit f4c19ac9c24c5c5dcab4e961e4d7f8f5709c650e upstream

There are new types and helpers that are supposed to be used in
new code.

As a preparation to get rid of legacy types and API functions do
the conversion here.

The conversion fixes a potential bug in int340x_thermal as well
since we have to use memcmp() on binary data.

Acked-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>